+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.3.0] - 2021-07-23
+### Changed 
+- Switch to Moulinette Cloud (new servers)
+### Added
+- Claim a gift from a creator
+
 ## [3.2.1] - 2021-07-11
 ### Added
 - Patreon : disconnect & reconnect buttons